[proliferate] consolidate cowork mcp server
