+++ anyharness/crates/anyharness-lib/src/domains/cowork/delegation/mod.rs
@@ -1,3 +1,2 @@
-pub mod mcp;
 pub mod model;
 pub mod service;
