+++ anyharness/crates/anyharness-lib/src/domains/cowork/mcp_server/auth.rs
